refactor: update plugin dependency management, remove OpenResty, and expand database and PHP version support.

+++ plugins/mariadb/versions/10.6/install.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:/opt/homebrew/bin:~/bin
+PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin:/opt/homebrew/bin
 export PATH
 
 # https://mariadb.org/download/?t=mariadb
@@ -7,26 +7,17 @@ export PATH
 curPath=`pwd`
 rootPath=$(dirname "$curPath")
 rootPath=$(dirname "$rootPath")
-serverPath=$(dirname "$rootPath")/server
+serverPath=$(dirname "$rootPath")
 sysName=`uname`
 
-install_tmp=${rootPath}/tmp/slemp_install.pl
 mariadbDir=${serverPath}/source/mariadb
 
-MY_VER=10.6.12
+MY_VER=10.6.21
 
 Install_app()
 {
 	mkdir -p ${mariadbDir}
-	echo 'Installing script file...' > $install_tmp
-
-	if id mysql &> /dev/null ;then
-	    echo "mysql uid is `id -u www`"
-	    echo "mysql shell is `grep "^www:" /etc/passwd |cut -d':' -f7 `"
-	else
-	    groupadd mysql
-		useradd -g mysql mysql
-	fi
+	echo '正在安装脚本文件...'
 
 	if [ "$sysName" != "Darwin" ];then
 		mkdir -p /var/log/mariadb
@@ -61,7 +52,7 @@ Install_app()
 	# https://mirrors.aliyun.com/mariadb/mariadb-10.8.3/source/mariadb-10.8.3.tar.gz
 	# if [ ! -f ${mariadbDir}/mariadb-${MY_VER}.tar.gz ];then
 	# 	wget --no-check-certificate -O ${mariadbDir}/mariadb-${MY_VER}.tar.gz --tries=3 https://mirrors.aliyun.com/mariadb/mariadb-${MY_VER}/source/mariadb-${MY_VER}.tar.gz
-
+		
 	# fi
 
 	# https://downloads.mariadb.org/interstitial/mariadb-10.6.8/source/mariadb-10.6.8.tar.gz
@@ -72,7 +63,7 @@ Install_app()
 	if [ ! -d ${mariadbDir}/mariadb-${MY_VER} ];then
 		 cd ${mariadbDir} && tar -zxvf  ${mariadbDir}/mariadb-${MY_VER}.tar.gz
 	fi
-
+	
 	OPTIONS=''
 	if [ "$sysName" == "Darwin" ];then
 		OPTIONS='-DPLUGIN_TOKUDB=NO'
@@ -105,22 +96,23 @@ Install_app()
 
 		if [ -d $serverPath/mariadb ];then
 			echo '10.6' > $serverPath/mariadb/version.pl
-			echo 'The installation is complete' > $install_tmp
+			echo '安装完成'
 		else
-			echo 'Installation failed' > $install_tmp
+			echo '安装失败'
 			echo 'install fail'>&2
 			exit 1
 		fi
 	fi
 
-	rm -rf ${mariadbDir}/mariadb-${MY_VER}
-	rm -rf ${mariadbDir}/mariadb-${MY_VER}.tar.gz
+	if [ -d ${mariadbDir}/mariadb-${MY_VER} ];then
+		rm -rf ${mariadbDir}/mariadb-${MY_VER}
+	fi
 }
 
 Uninstall_app()
 {
 	rm -rf $serverPath/mariadb
-	echo 'Uninstall complete' > $install_tmp
+	echo '卸载完成'
 }
 
 action=$1
